feat: image class 패키지 변경 및 validation
