--- core/src/main/java/kr/co/toppings/core/global/entity/embedded/Image.java
@@ -1,23 +1,0 @@
-package kr.co.toppings.core.global.entity.embedded;
-
-import javax.persistence.Embeddable;
-
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-@Embeddable
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Image {
-
-	private String url;
-
-	private String path;
-
-	public Image(
-		String url,
-		String path
-	) {
-		this.url = url;
-		this.path = path;
-	}
-}
